requirements.txt
fastapi
uvicorn
redis
scikit-learn
pandas
joblib
python-dotenv
sqlalchemy
asyncpg
alembic
